refactor training code into translation and representation modules
