+++ applications/contrastive_phenotyping/contrastive_scripts/demo_fit.py
@@ -3,7 +3,7 @@ from lightning.pytorch.callbacks import ModelCheckpoint
 from lightning.pytorch.loggers import TensorBoardLogger
 
 from viscy.data.triplet import TripletDataModule
-from viscy.light.engine import ContrastiveModule
+from viscy.representation.engine import ContrastiveModule
 
 
 def main():
