+++ examples/virtual_staining/VS_model_inference/demo_vsneuromast.py
@@ -13,9 +13,9 @@ from iohub import open_ome_zarr
 from plot import plot_vs_n_fluor
 from viscy.data.hcs import HCSDataModule
 # Viscy classes for the trainer and model
-from viscy.light.engine import VSUNet
-from viscy.light.predict_writer import HCSPredictionWriter
-from viscy.light.trainer import VSTrainer
+from viscy.translation.engine import VSUNet
+from viscy.translation.predict_writer import HCSPredictionWriter
+from viscy.translation.trainer import VSTrainer
 from viscy.transforms import NormalizeSampled
 
 # %% [markdown]
