+++ examples/virtual_staining/dlmbl_exercise/solution.py
@@ -118,8 +118,8 @@ from tqdm import tqdm
 from viscy.data.hcs import HCSDataModule
 from viscy.evaluation.evaluation_metrics import mean_average_precision
 # Trainer class and UNet.
-from viscy.light.engine import MixedLoss, VSUNet
-from viscy.light.trainer import VSTrainer
+from viscy.translation.engine import MixedLoss, VSUNet
+from viscy.translation.trainer import VSTrainer
 # training augmentations
 from viscy.transforms import (NormalizeSampled, RandAdjustContrastd,
                               RandAffined, RandGaussianNoised,
