+++ viscy/representation/embedding_writer.py
